requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# LLM Integration
anthropic>=0.25.0
openai>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0